hwc: Stop queuing compositions to disabled default display
When HDMI is the default screen it may not be attached even
though the UI is active. UI compositions will keep coming but
the DSS manager will have problems if being programmed with
content incompatible with the intermediate display mode.

The last_mode variable is reset to "0" when hdmi panel is
not present. The on_tv variable indicates hdmi is default
panel.  So if both are set then we need to make the overlay
number in composition to be "0".

OMAPLFB will skip posting the composition to dss driver when
overlay as "0".

There are still possibly compositions in the DSSCOMP queue
or in flight from PVR services to OMAPLFB so this behavior
doesn't cover all cases - so this should be considered a
work-around.

kernel dependency:
http://review.omapzoom.org/#/c/26982/

Change-Id: I1eae6aad869c2918274b3eaa1b7370e491a15f3f
Signed-off-by: Dandawate Saket <dsaket@ti.com>
Signed-off-by: Tony Lofthouse <tony.lofthouse@ti.com>